--- src/main/scala/libra/dimensions/All.scala
@@ -1,6 +1,0 @@
-package libra
-package dimensions
-
-trait All extends AngleDimension
-    with TimeDimension
-    with MassDimension
