+++ src/main/scala/libra/implicits/force.scala
@@ -1,0 +1,9 @@
+package libra
+package implicits
+
+trait ForceImplicits {
+
+  implicit final class ForceQuantityOps[A](val a: A) {
+    def N: ForceQuantity[A, Kilogram, Metre, Second] = Quantity(a)
+  }
+}
