--- src/main/scala/libra/si/MetricUnit.scala
@@ -1,6 +1,0 @@
-package libra
-package si
-
-import singleton.ops._
-
-trait MetricUnit[I <: XInt, D] extends UnitOfMeasure[D]
