+++ src/main/scala/libra/units/AmountUnits.scala
@@ -1,12 +1,12 @@
 package libra
-package dimensions
+package units
 
 import ops.base.{Show, ConversionFactor}
 import spire._, spire.algebra._, spire.math._, spire.implicits._
 import singleton.ops._
 import shapeless._
 
-trait AmountDimension {
+trait AmountUnits {
 
   type Amount
