+++ src/main/scala/libra/units/TimeUnits.scala
@@ -1,12 +1,12 @@
 package libra
-package dimensions
+package units
 
 import ops.base.{Show, ConversionFactor}
 import spire._, spire.algebra._, spire.math._, spire.implicits._
 import singleton.ops._
 import shapeless._
 
-trait TimeDimension extends MetricUnitImplicits {
+trait TimeUnits extends MetricUnitImplicits {
 
   type Time
