+++ src/test/scala/libra/QuantitySpec.scala
@@ -9,7 +9,7 @@ import singleton.ops.impl._
 
 import spire._, spire.algebra._, spire.math._, spire.implicits._
 
-import libra.ops.quantity._, libra.si._
+import libra.ops.quantity._, libra.implicits._
 
 class QuantitySpec extends FlatSpec {
 
@@ -115,4 +115,3 @@ class QuantitySpec extends FlatSpec {
   }
 
 }
-
