+++ src/test/scala/libra/ops/DimensionsSpec.scala
@@ -1,10 +1,11 @@
 package libra
+package ops
 
 import org.scalatest._
 
 import shapeless._
 import shapeless.test._
-import libra.ops.dimensions._, libra.si._
+import libra.ops.dimensions._, libra.implicits._
 import spire.implicits._
 
 class DimensionsSpec extends FlatSpec {
